refactor: extract duplicated timeAgo and formatPrice to shared utils
Four components had their own versions of price formatting and
relative time display. Consolidated into src/utils.ts and updated
all imports. No behavior change.

Closes #38
